finished lottery scheduler etc

+++ xv6-riscv/kernel/defs.h
@@ -107,7 +107,7 @@ void            procdump(void);
 void		    process_count_print(void);
 void		    syscall_count_print(void);
 void            mem_pages_count_print(void);
-void		    set_tickets(int); // for lottery scheduling
+void		    set_tickets(uint64);
 void            sched_statistics(void);
 
 // swtch.S
